homework 4
