+++ js/image/values.js
@@ -1,0 +1,33 @@
+export const descriptions = [
+  'На этом фото запечатлена величественная красота зимнего леса.',
+  'Солнечные лучи пробиваются сквозь листву, создавая волшебный узор.',
+  'Озеро отражает небо, словно огромное зеркало.',
+  'Уютный уголок с камином и креслом, идеален для чтения.',
+  'Закат окрасил небо в потрясающие оттенки оранжевого и розового.',
+  'Горы под снежной шапкой выглядят величественно и неприступно.',
+  'Река спокойно течет между зелеными берегами, создавая атмосферу спокойствия.',
+  'Город засыпает, погружаясь в мягкий свет уличных фонарей.'
+];
+export const commentDescriptions = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+export const commentatorNames = [
+  'Иван',
+  'Петр',
+  'Сергей',
+  'Анна',
+  'Мария',
+  'Екатерина',
+  'Наталья',
+  'Владимир',
+  'Илья',
+  'Александр',
+  'Андрей',
+  'Виктор',
+  'Антон',
+];
